Rename parcel-link util to utils
utils is the convention in the monorepo
